+++ src/utils/c_utils/pseudoinverse.h
@@ -2,6 +2,6 @@
 #define PSEUDOINVERSE_H
 
 // Function declaration
-void computePseudoinverse(double *A, double *A_inv, int m, int n);
+void computePseudoinverse(float *A, float *A_inv, int m, int n);
 
 #endif // PSEUDOINVERSE_H
