+++ src/utils/c_utils/utils.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 // Function to print a matrix
-void printMatrix(double *A, int m, int n) {
+void printMatrix(float *A, int m, int n) {
     for (int i = 0; i < m; i++) {
         for (int j = 0; j < n; j++) {
             printf("%f ", A[i * n + j]);
